refactor: move functions to independent files + cleanup
