+++ lib/entities/board.js
