+++ lib/entities/group.js
