+++ lib/index.js
@@ -1,0 +1,11 @@
+const createPopulation = require('./createPopulation');
+const select = require('./select');
+const mutate = require('./mutate');
+const cross = require('./cross');
+
+module.exports =  {
+    createPopulation,
+    select,
+    mutate,
+    cross,
+};
